+++ AzurLaneAPI/Controllers/ScrapersController.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using AzurLaneClasses;
 using Microsoft.AspNetCore.Mvc;
 
@@ -10,6 +13,12 @@ namespace AzurLaneAPI.Controllers
 		public ScrapersController(AzurLaneDbContext context)
 		{
 			_context = context;
+		}
+
+		[HttpGet(Routes.V1.Routes.Scrapers.Ships.GetPageUrls)]
+		public async Task<ActionResult<List<String>>> ScrapeShips()
+		{
+			return Scrapers.ShipsScraper.GetShipWikiUrls();
 		}
 	}
 }
