+++ AzurLaneAPI/Routes/V1/Routes.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace AzurLaneAPI.Routes.V1
 {
-	public static class Routes
+	public static partial class Routes
 	{
 		public const String Base = "/v1";
 
@@ -79,142 +79,6 @@ namespace AzurLaneAPI.Routes.V1
 			public const String GetPool = "/construction/{id}";
 		}
 
-		public static class Equipment
-		{
-			public static class DestroyerGuns
-			{
-				public const String GetAll = "/equipment/destroyerguns";
-				public const String GetId = "/equipment/destroyerguns/{id}";
-				public const String Create = "/equipment/destroyerguns";
-				public const String Update = "/equipment/destroyerguns/{id}";
-				public const String Delete = "/equipment/destroyerguns/{id}";
-			}
-
-			public static class LightCruiserGuns
-			{
-				public const String GetAll = "/equipment/lightcruiserguns";
-				public const String GetId = "/equipment/lightcruiserguns/{id}";
-				public const String Create = "/equipment/lightcruiserguns";
-				public const String Update = "/equipment/lightcruiserguns/{id}";
-				public const String Delete = "/equipment/lightcruiserguns/{id}";
-			}
-
-			public static class HeavyCruiserGuns
-			{
-				public const String GetAll = "/equipment/heavycruiserguns";
-				public const String GetId = "/equipment/heavycruiserguns/{id}";
-				public const String Create = "/equipment/heavycruiserguns";
-				public const String Update = "/equipment/heavycruiserguns/{id}";
-				public const String Delete = "/equipment/heavycruiserguns/{id}";
-			}
-
-			public static class LargeCruiserGuns
-			{
-				public const String GetAll = "/equipment/largecruiserguns";
-				public const String GetId = "/equipment/largecruiserguns/{id}";
-				public const String Create = "/equipment/largecruiserguns";
-				public const String Update = "/equipment/largecruiserguns/{id}";
-				public const String Delete = "/equipment/largecruiserguns/{id}";
-			}
-
-			public static class BattleshipGuns
-			{
-				public const String GetAll = "/equipment/battleshipguns";
-				public const String GetId = "/equipment/battleshipguns/{id}";
-				public const String Create = "/equipment/battleshipguns";
-				public const String Update = "/equipment/battleshipguns/{id}";
-				public const String Delete = "/equipment/battleshipguns/{id}";
-			}
-
-			public static class Torpedoes
-			{
-				public const String GetAll = "/equipment/torpedoes";
-				public const String GetId = "/equipment/torpedoes/{id}";
-				public const String Create = "/equipment/torpedoes";
-				public const String Update = "/equipment/torpedoes/{id}";
-				public const String Delete = "/equipment/torpedoes/{id}";
-			}
-
-			public static class SubmarineTorpedoes
-			{
-				public const String GetAll = "/equipment/submarinetorpedoes";
-				public const String GetId = "/equipment/submarinetorpedoes/{id}";
-				public const String Create = "/equipment/submarinetorpedoes";
-				public const String Update = "/equipment/submarinetorpedoes/{id}";
-				public const String Delete = "/equipment/submarinetorpedoes/{id}";
-			}
-
-			public static class FighterPlanes
-			{
-				public const String GetAll = "/equipment/fighterplanes";
-				public const String GetId = "/equipment/fighterplanes/{id}";
-				public const String Create = "/equipment/fighterplanes";
-				public const String Update = "/equipment/fighterplanes/{id}";
-				public const String Delete = "/equipment/fighterplanes/{id}";
-			}
-
-			public static class DiveBomberPlanes
-			{
-				public const String GetAll = "/equipment/divebomberplanes";
-				public const String GetId = "/equipment/divebomberplanes/{id}";
-				public const String Create = "/equipment/divebomberplanes";
-				public const String Update = "/equipment/divebomberplanes/{id}";
-				public const String Delete = "/equipment/divebomberplanes/{id}";
-			}
-
-			public static class TorpedoBomberPlanes
-			{
-				public const String GetAll = "/equipment/torpedobomberplanes";
-				public const String GetId = "/equipment/torpedobomberplanes/{id}";
-				public const String Create = "/equipment/torpedobomberplanes";
-				public const String Update = "/equipment/torpedobomberplanes/{id}";
-				public const String Delete = "/equipment/torpedobomberplanes/{id}";
-			}
-
-			public static class Seaplanes
-			{
-				public const String GetAll = "/equipment/seaplanes";
-				public const String GetId = "/equipment/seaplanes/{id}";
-				public const String Create = "/equipment/seaplanes";
-				public const String Update = "/equipment/seaplanes/{id}";
-				public const String Delete = "/equipment/seaplanes/{id}";
-			}
-
-			public static class AntiAirGuns
-			{
-				public const String GetAll = "/equipment/antiairguns";
-				public const String GetId = "/equipment/antiairguns/{id}";
-				public const String Create = "/equipment/antiairguns";
-				public const String Update = "/equipment/antiairguns/{id}";
-				public const String Delete = "/equipment/antiairguns/{id}";
-			}
-
-			public static class Auxiliary
-			{
-				public const String GetAll = "/equipment/auxiliary";
-				public const String GetId = "/equipment/auxiliary/{id}";
-				public const String Create = "/equipment/auxiliary";
-				public const String Update = "/equipment/auxiliary/{id}";
-				public const String Delete = "/equipment/auxiliary/{id}";
-			}
-
-			public static class Cargo
-			{
-				public const String GetAll = "/equipment/cargo";
-				public const String GetId = "/equipment/cargo/{id}";
-				public const String Create = "/equipment/cargo";
-				public const String Update = "/equipment/cargo/{id}";
-				public const String Delete = "/equipment/cargo/{id}";
-			}
-
-			public static class AntiSubmarine
-			{
-				public const String GetAll = "/equipment/antisubmarine";
-				public const String GetId = "/equipment/antisubmarine/{id}";
-				public const String Create = "/equipment/antisubmarine";
-				public const String Update = "/equipment/antisubmarine/{id}";
-				public const String Delete = "/equipment/antisubmarine/{id}";
-			}
-		}
+		
 	}
 }
