+++ AzurLaneAPI/Routes/V1/ScraperRoutes.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace AzurLaneAPI.Routes.V1
+{
+    public static partial class Routes
+    {
+        public static class Scrapers 
+		{
+			public static class Ships 
+			{
+				public const String GetPageUrls = "/scrapers/ships/urls";
+			}
+		}
+    }
+}
